{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "{{ cookiecutter.description }}"
readme = "README.md"
authors = [
    { name = "{{ cookiecutter.author_name }}", email = "{{ cookiecutter.email }}" }
]
requires-python = ">=3.13"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.7",
    "django-allauth[socialaccount]>=65.12.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "drf-yasg==1.21.7",
    "setuptools>=80.9.0",
    "black>=25.9.0",
    "pre-commit>=4.4.0",
]

